RescMeal V1.2

+++ Backend/src/main/java/dev/group21/rescmeal/services/BusinessService.java
@@ -1,28 +1,26 @@
 package dev.group21.rescmeal.services;
 
 import dev.group21.rescmeal.model.Business;
-import dev.group21.rescmeal.model.BusinessPhoto;
-import dev.group21.rescmeal.repository.BusinessPhotoRepository;
 import dev.group21.rescmeal.repository.BusinessRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import java.io.File;
+import java.util.Optional;
 
 @Service
 @Transactional
 public class BusinessService {
     private final BusinessRepository businessRepository;
-    private final BusinessPhotoRepository businessPhotoRepository;
 
     /**
      * Init methods for Business Service.
      * @param businessRepository Business Repository.
-     * @param businessPhotoRepository BusinessPhoto Repository.
      */
-    public BusinessService(BusinessRepository businessRepository, BusinessPhotoRepository businessPhotoRepository) {
+    public BusinessService(BusinessRepository businessRepository) {
         this.businessRepository = businessRepository;
-        this.businessPhotoRepository = businessPhotoRepository;
     }
 
     /**
@@ -31,14 +29,7 @@ public class BusinessService {
      * @return Business entity
      */
     public Business createBusiness(Business business) {
-        // TODO Prevent infinite recursion on creation, and a proper creation of Photo and Business.
-        BusinessPhoto businessPhoto = business.getBusinessPhoto();
-        business.setBusinessPhoto(null);
-        Business newBusiness = businessRepository.save(business);
-        businessPhoto.setBusiness_id(newBusiness.getId());
-        businessPhotoRepository.save(businessPhoto);
-
-        return newBusiness;
+        return businessRepository.save(business);
     }
 
     /**
@@ -47,7 +38,6 @@ public class BusinessService {
      * @return Business entity
      */
     public Business updateBusiness(Business updateBusiness) {
-        businessPhotoRepository.saveAndFlush(updateBusiness.getBusinessPhoto());
         return businessRepository.saveAndFlush(updateBusiness);
     }
 
@@ -59,34 +49,42 @@ public class BusinessService {
      */
     public Business dynamicUpdateBusiness(Business existingBusiness, Business updateBusiness) {
         // Update fields that were sent with some value.
-        if(updateBusiness.getBusiness_name() == null) updateBusiness.setBusiness_name(existingBusiness.getBusiness_name());
-        if(updateBusiness.getBusiness_type() == null) updateBusiness.setBusiness_type(existingBusiness.getBusiness_type());
+        if(updateBusiness.getName() == null) updateBusiness.setName(existingBusiness.getName());
+        if(updateBusiness.getType() == null) updateBusiness.setType(existingBusiness.getType());
         if(updateBusiness.getAddress() == null) updateBusiness.setAddress(existingBusiness.getAddress());
-        if(updateBusiness.getEmail() == null) updateBusiness.setEmail(existingBusiness.getEmail());
-        if(updateBusiness.getPassword() == null) updateBusiness.setPassword(existingBusiness.getPassword());
         if(updateBusiness.getPhone() == null) updateBusiness.setPhone(existingBusiness.getPhone());
-        if(updateBusiness.getBusiness_time() == null) updateBusiness.setBusiness_time(existingBusiness.getBusiness_time());
+        if(updateBusiness.getSchedule() == null) updateBusiness.setSchedule(existingBusiness.getSchedule());
         if(updateBusiness.getCvu() == null) updateBusiness.setCvu(existingBusiness.getCvu());
-        // Update photo
-        if(updateBusiness.getBusinessPhoto() != null) businessPhotoRepository.saveAndFlush(updateBusiness.getBusinessPhoto());
+        if(updateBusiness.getImage() == null) updateBusiness.setImage(existingBusiness.getImage());
         return businessRepository.saveAndFlush(updateBusiness);
     }
 
     /**
-     * Delete a Business entity by id and its BusinessPhoto relation.
-     * @param id Integer id of Business entity.
+     * Delete a business and their image from the database.
+     * @param id Long id of Business entity.
      */
-    public void deleteBusiness(Integer id) {
-        businessPhotoRepository.deleteById(id);
-        businessRepository.deleteById(id);
+    public void deleteBusiness(Long id) {
+        Business business = businessRepository.findById(id).orElse(null);
+        if (business != null) {
+            // Ruta al directorio public de tu proyecto Next.js
+            String publicDir = System.getProperty("user.dir") + "/../Frontend/public/Business/";
+            String imagePath = publicDir + business.getImage();
+            // Eliminar la imagen si existe
+            File imageFile = new File(imagePath);
+            if (imageFile.exists()) {
+                imageFile.delete();
+            }
+            // Eliminar el negocio de la base de datos
+            businessRepository.deleteById(id);
+        }
     }
 
     /**
      * Get method for Business entity, finds by entity id.
-     * @param id Integer id or Business entity.
+     * @param id Long id of Business entity.
      * @return Business entity.
      */
-    public Business getBusiness(Integer id) {
+    public Business getBusiness(Long id) {
         return businessRepository.findById(id).orElse(null);
     }
 
@@ -94,7 +92,9 @@ public class BusinessService {
      * List all Business entity present in database.
      * @return List of Business entities.
      */
-    public List<Business> getAllBusiness() {
-        return businessRepository.findAll();
+    public Page<Business> getAllBusiness(Pageable parameters) {
+        return businessRepository.findAll(parameters);
     }
+
+    public Optional<Business> getBusinessById(Long id) { return businessRepository.findById(id); }
 }
